feat(front-kit-3): moved Form to front-kit
