+++ packages/falcon-front-kit/src/Forms/index.ts
@@ -1,4 +1,5 @@
 export * from './FormContext';
+export * from './Form';
 export * from './Field';
 export * from './validators';
 export * from './inputTypeToDefaultValidatorsMapper';
